refactor(sequencer): move asset state methods to asset module
