+++ crates/astria-sequencer/src/accounts/mod.rs
@@ -1,4 +1,9 @@
 pub(crate) mod action;
 pub(crate) mod component;
 pub(crate) mod query;
-pub(crate) mod state_ext;
+mod state_ext;
+
+pub(crate) use state_ext::{
+    StateReadExt,
+    StateWriteExt,
+};
