+++ crates/astria-sequencer/src/authority/state_ext.rs
@@ -5,13 +5,9 @@ use anyhow::{
     Context,
     Result,
 };
-use astria_core::{
-    crypto::VerificationKey,
-    primitive::v1::{
-        Address,
-        ADDRESS_LEN,
-    },
-    protocol::transaction::v1alpha1::action::ValidatorUpdate,
+use astria_core::primitive::v1::{
+    Address,
+    ADDRESS_LEN,
 };
 use async_trait::async_trait;
 use borsh::{
@@ -22,84 +18,13 @@ use cnidarium::{
     StateRead,
     StateWrite,
 };
-use serde::{
-    Deserialize,
-    Serialize,
-};
 use tracing::instrument;
+
+use super::ValidatorSet;
 
 /// Newtype wrapper to read and write an address from rocksdb.
 #[derive(BorshSerialize, BorshDeserialize, Debug)]
 struct SudoAddress([u8; ADDRESS_LEN]);
-
-/// Newtype wrapper to read and write a validator set or set of updates from rocksdb.
-///
-/// Contains a map of hex-encoded public keys to validator updates.
-#[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
-pub(crate) struct ValidatorSet(BTreeMap<ValidatorSetKey, ValidatorUpdate>);
-
-#[derive(Copy, Clone, Debug, Serialize, Deserialize, PartialEq, Eq, Ord, PartialOrd)]
-pub(crate) struct ValidatorSetKey(#[serde(with = "::hex::serde")] [u8; ADDRESS_LEN]);
-
-impl From<[u8; ADDRESS_LEN]> for ValidatorSetKey {
-    fn from(value: [u8; ADDRESS_LEN]) -> Self {
-        Self(value)
-    }
-}
-
-impl From<VerificationKey> for ValidatorSetKey {
-    fn from(value: VerificationKey) -> Self {
-        Self(value.address_bytes())
-    }
-}
-
-impl ValidatorSet {
-    pub(crate) fn new_from_updates(updates: Vec<ValidatorUpdate>) -> Self {
-        Self(
-            updates
-                .into_iter()
-                .map(|update| (update.verification_key.into(), update))
-                .collect::<BTreeMap<_, _>>(),
-        )
-    }
-
-    pub(crate) fn len(&self) -> usize {
-        self.0.len()
-    }
-
-    pub(crate) fn get<T: Into<ValidatorSetKey>>(&self, address: T) -> Option<&ValidatorUpdate> {
-        self.0.get(&address.into())
-    }
-
-    pub(crate) fn push_update(&mut self, update: ValidatorUpdate) {
-        self.0.insert(update.verification_key.into(), update);
-    }
-
-    pub(crate) fn remove<T: Into<ValidatorSetKey>>(&mut self, address: T) {
-        self.0.remove(&address.into());
-    }
-
-    /// Apply updates to the validator set.
-    ///
-    /// If the power of a validator is set to 0, remove it from the set.
-    /// Otherwise, update the validator's power.
-    pub(crate) fn apply_updates(&mut self, validator_updates: ValidatorSet) {
-        for (address, update) in validator_updates.0 {
-            match update.power {
-                0 => self.0.remove(&address),
-                _ => self.0.insert(address, update),
-            };
-        }
-    }
-
-    pub(crate) fn try_into_cometbft(self) -> anyhow::Result<Vec<tendermint::validator::Update>> {
-        self.0
-            .into_values()
-            .map(crate::utils::sequencer_to_cometbft_validator)
-            .collect::<Result<Vec<_>, _>>()
-            .context("failed to map one or more astria validators to cometbft validators")
-    }
-}
 
 const SUDO_STORAGE_KEY: &str = "sudo";
 const VALIDATOR_SET_STORAGE_KEY: &str = "valset";
@@ -204,9 +129,11 @@ mod test {
     use super::{
         StateReadExt as _,
         StateWriteExt as _,
-        ValidatorSet,
     };
-    use crate::test_utils::verification_key;
+    use crate::{
+        authority::ValidatorSet,
+        test_utils::verification_key,
+    };
 
     fn empty_validator_set() -> ValidatorSet {
         ValidatorSet::new_from_updates(vec![])
