+++ crates/astria-sequencer/src/bridge/mod.rs
@@ -4,6 +4,10 @@ mod bridge_unlock_action;
 pub(crate) mod component;
 pub(crate) mod init_bridge_account_action;
 pub(crate) mod query;
-pub(crate) mod state_ext;
+mod state_ext;
 
 pub(crate) use bridge_lock_action::get_deposit_byte_len;
+pub(crate) use state_ext::{
+    StateReadExt,
+    StateWriteExt,
+};
