+++ crates/astria-sequencer/src/ibc/mod.rs
@@ -3,4 +3,9 @@ pub(crate) mod host_interface;
 pub(crate) mod ibc_relayer_change;
 pub(crate) mod ics20_transfer;
 pub(crate) mod ics20_withdrawal;
-pub(crate) mod state_ext;
+mod state_ext;
+
+pub(crate) use state_ext::{
+    StateReadExt,
+    StateWriteExt,
+};
