+++ crates/astria-sequencer/src/sequence/mod.rs
@@ -1,5 +1,9 @@
 pub(crate) mod action;
 pub(crate) mod component;
-pub(crate) mod state_ext;
+mod state_ext;
 
 pub(crate) use action::calculate_fee_from_state;
+pub(crate) use state_ext::{
+    StateReadExt,
+    StateWriteExt,
+};
