+++ crates/astria-sequencer/src/state_ext.rs
@@ -3,58 +3,18 @@ use anyhow::{
     Context as _,
     Result,
 };
-use astria_core::primitive::v1::asset;
 use async_trait::async_trait;
 use cnidarium::{
     StateRead,
     StateWrite,
 };
-use futures::StreamExt as _;
-use tendermint::{
-    abci::{
-        Event,
-        EventAttributeIndexExt as _,
-    },
-    Time,
-};
+use tendermint::Time;
 use tracing::instrument;
 
-const NATIVE_ASSET_KEY: &[u8] = b"nativeasset";
 const REVISION_NUMBER_KEY: &str = "revision_number";
-const BLOCK_FEES_PREFIX: &str = "block_fees/";
-const FEE_ASSET_PREFIX: &str = "fee_asset/";
 
 fn storage_version_by_height_key(height: u64) -> Vec<u8> {
     format!("storage_version/{height}").into()
-}
-
-fn block_fees_key<TAsset: Into<asset::IbcPrefixed>>(asset: TAsset) -> String {
-    format!(
-        "{BLOCK_FEES_PREFIX}{}",
-        crate::storage_keys::hunks::Asset::from(asset)
-    )
-}
-
-fn fee_asset_key<TAsset: Into<asset::IbcPrefixed>>(asset: TAsset) -> String {
-    format!(
-        "{FEE_ASSET_PREFIX}{}",
-        crate::storage_keys::hunks::Asset::from(asset)
-    )
-}
-
-/// Creates `abci::Event` of kind `tx.fees` for sequencer fee reporting
-fn construct_tx_fee_event<TAsset>(asset: &TAsset, fee_amount: u128, action_type: String) -> Event
-where
-    TAsset: Into<asset::IbcPrefixed> + std::fmt::Display + Send,
-{
-    Event::new(
-        "tx.fees",
-        [
-            ("asset", asset.to_string()).index(),
-            ("feeAmount", fee_amount.to_string()).index(),
-            ("actionType", action_type).index(),
-        ],
-    )
 }
 
 #[async_trait]
@@ -139,82 +99,6 @@ pub(crate) trait StateReadExt: StateRead {
         };
         Ok(u64::from_be_bytes(bytes))
     }
-
-    #[instrument(skip_all)]
-    async fn get_native_asset_denom(&self) -> Result<String> {
-        let Some(bytes) = self
-            .nonverifiable_get_raw(NATIVE_ASSET_KEY)
-            .await
-            .context("failed to read raw native_asset_denom from state")?
-        else {
-            bail!("native asset denom not found");
-        };
-
-        String::from_utf8(bytes).context("failed to parse native asset denom from raw bytes")
-    }
-
-    #[instrument(skip_all)]
-    async fn get_block_fees(&self) -> Result<Vec<(asset::IbcPrefixed, u128)>> {
-        // let mut fees: Vec<(asset::Id, u128)> = Vec::new();
-        let mut fees = Vec::new();
-
-        let mut stream =
-            std::pin::pin!(self.nonverifiable_prefix_raw(BLOCK_FEES_PREFIX.as_bytes()));
-        while let Some(Ok((key, value))) = stream.next().await {
-            // if the key isn't of the form `block_fees/{asset_id}`, then we have a bug
-            // in `put_block_fees`
-            let suffix = key
-                .strip_prefix(BLOCK_FEES_PREFIX.as_bytes())
-                .expect("prefix must always be present");
-            let asset = std::str::from_utf8(suffix)
-                .context("key suffix was not utf8 encoded; this should not happen")?
-                .parse::<crate::storage_keys::hunks::Asset>()
-                .context("failed to parse storage key suffix as address hunk")?
-                .get();
-
-            let Ok(bytes): Result<[u8; 16], _> = value.try_into() else {
-                bail!("failed turning raw block fees bytes into u128; not 16 bytes?");
-            };
-
-            fees.push((asset, u128::from_be_bytes(bytes)));
-        }
-
-        Ok(fees)
-    }
-
-    #[instrument(skip_all)]
-    async fn is_allowed_fee_asset<TAsset>(&self, asset: TAsset) -> Result<bool>
-    where
-        TAsset: Into<asset::IbcPrefixed> + std::fmt::Display + Send,
-    {
-        Ok(self
-            .nonverifiable_get_raw(fee_asset_key(asset).as_bytes())
-            .await
-            .context("failed to read raw fee asset from state")?
-            .is_some())
-    }
-
-    #[instrument(skip_all)]
-    async fn get_allowed_fee_assets(&self) -> Result<Vec<asset::IbcPrefixed>> {
-        let mut assets = Vec::new();
-
-        let mut stream = std::pin::pin!(self.nonverifiable_prefix_raw(FEE_ASSET_PREFIX.as_bytes()));
-        while let Some(Ok((key, _))) = stream.next().await {
-            // if the key isn't of the form `fee_asset/{asset_id}`, then we have a bug
-            // in `put_allowed_fee_asset`
-            let suffix = key
-                .strip_prefix(FEE_ASSET_PREFIX.as_bytes())
-                .expect("prefix must always be present");
-            let asset = std::str::from_utf8(suffix)
-                .context("key suffix was not utf8 encoded; this should not happen")?
-                .parse::<crate::storage_keys::hunks::Asset>()
-                .context("failed to parse storage key suffix as address hunk")?
-                .get();
-            assets.push(asset);
-        }
-
-        Ok(assets)
-    }
 }
 
 impl<T: StateRead> StateReadExt for T {}
@@ -253,74 +137,6 @@ pub(crate) trait StateWriteExt: StateWrite {
             version.to_be_bytes().to_vec(),
         );
     }
-
-    #[instrument(skip_all)]
-    fn put_native_asset_denom(&mut self, denom: &str) {
-        self.nonverifiable_put_raw(NATIVE_ASSET_KEY.to_vec(), denom.as_bytes().to_vec());
-    }
-
-    /// Adds `amount` to the block fees for `asset`.
-    #[instrument(skip_all)]
-    async fn get_and_increase_block_fees<TAsset>(
-        &mut self,
-        asset: TAsset,
-        fee_amount: u128,
-        action_type: String,
-    ) -> Result<()>
-    where
-        TAsset: Into<asset::IbcPrefixed> + std::fmt::Display + Send + Clone,
-    {
-        let block_fees_key = block_fees_key(asset.clone());
-        let current_amount = self
-            .nonverifiable_get_raw(block_fees_key.as_bytes())
-            .await
-            .context("failed to read raw block fees from state")?
-            .map(|bytes| {
-                let Ok(bytes): Result<[u8; 16], _> = bytes.try_into() else {
-                    // this shouldn't happen
-                    bail!("failed turning raw block fees bytes into u128; not 16 bytes?");
-                };
-                Ok(u128::from_be_bytes(bytes))
-            })
-            .transpose()?
-            .unwrap_or_default();
-
-        let new_amount = current_amount
-            .checked_add(fee_amount)
-            .context("block fees overflowed u128")?;
-        self.nonverifiable_put_raw(block_fees_key.into(), new_amount.to_be_bytes().to_vec());
-
-        // record the fee event to the state cache
-        let tx_fee_event = construct_tx_fee_event(&asset, fee_amount, action_type);
-        self.record(tx_fee_event);
-
-        Ok(())
-    }
-
-    #[instrument(skip_all)]
-    async fn clear_block_fees(&mut self) {
-        let mut stream =
-            std::pin::pin!(self.nonverifiable_prefix_raw(BLOCK_FEES_PREFIX.as_bytes()));
-        while let Some(Ok((key, _))) = stream.next().await {
-            self.nonverifiable_delete(key);
-        }
-    }
-
-    #[instrument(skip_all)]
-    fn put_allowed_fee_asset<TAsset>(&mut self, asset: TAsset)
-    where
-        TAsset: Into<asset::IbcPrefixed> + std::fmt::Display + Send,
-    {
-        self.nonverifiable_put_raw(fee_asset_key(asset).into(), vec![]);
-    }
-
-    #[instrument(skip_all)]
-    fn delete_allowed_fee_asset<TAsset>(&mut self, asset: TAsset)
-    where
-        TAsset: Into<asset::IbcPrefixed> + std::fmt::Display,
-    {
-        self.nonverifiable_delete(fee_asset_key(asset).into());
-    }
 }
 
 impl<T: StateWrite> StateWriteExt for T {}
@@ -343,9 +159,7 @@ fn revision_number_from_chain_id(chain_id: &str) -> u64 {
 }
 
 #[cfg(test)]
-mod test {
-    use std::collections::HashSet;
-
+mod tests {
     use cnidarium::StateDelta;
     use tendermint::Time;
 
@@ -354,20 +168,6 @@ mod test {
         StateReadExt as _,
         StateWriteExt as _,
     };
-    use crate::state_ext::{
-        block_fees_key,
-        fee_asset_key,
-    };
-
-    fn asset_0() -> astria_core::primitive::v1::asset::Denom {
-        "asset_0".parse().unwrap()
-    }
-    fn asset_1() -> astria_core::primitive::v1::asset::Denom {
-        "asset_1".parse().unwrap()
-    }
-    fn asset_2() -> astria_core::primitive::v1::asset::Denom {
-        "asset_2".parse().unwrap()
-    }
 
     #[test]
     fn revision_number_from_chain_id_regex() {
@@ -599,236 +399,5 @@ mod test {
             storage_version_update,
             "original but updated storage version was not what was expected"
         );
-    }
-
-    #[tokio::test]
-    async fn native_asset_denom() {
-        let storage = cnidarium::TempStorage::new().await.unwrap();
-        let snapshot = storage.latest_snapshot();
-        let mut state = StateDelta::new(snapshot);
-
-        // doesn't exist at first
-        state
-            .get_native_asset_denom()
-            .await
-            .expect_err("no native asset denom should exist at first");
-
-        // can write
-        let denom_orig = "denom_orig";
-        state.put_native_asset_denom(denom_orig);
-        assert_eq!(
-            state.get_native_asset_denom().await.expect(
-                "a native asset denomination was written and must exist inside the database"
-            ),
-            denom_orig,
-            "stored native asset denomination was not what was expected"
-        );
-
-        // can write new value
-        let denom_update = "denom_update";
-        state.put_native_asset_denom(denom_update);
-        assert_eq!(
-            state.get_native_asset_denom().await.expect(
-                "a native asset denomination update was written and must exist inside the database"
-            ),
-            denom_update,
-            "updated native asset denomination was not what was expected"
-        );
-    }
-
-    #[tokio::test]
-    async fn block_fee_read_and_increase() {
-        let storage = cnidarium::TempStorage::new().await.unwrap();
-        let snapshot = storage.latest_snapshot();
-        let mut state = StateDelta::new(snapshot);
-
-        // doesn't exist at first
-        let fee_balances_orig = state.get_block_fees().await.unwrap();
-        assert!(fee_balances_orig.is_empty());
-
-        // can write
-        let asset = asset_0();
-        let amount = 100u128;
-        state
-            .get_and_increase_block_fees(&asset, amount, "test".into())
-            .await
-            .unwrap();
-
-        // holds expected
-        let fee_balances_updated = state.get_block_fees().await.unwrap();
-        assert_eq!(
-            fee_balances_updated[0],
-            (asset.to_ibc_prefixed(), amount),
-            "fee balances are not what they were expected to be"
-        );
-    }
-
-    #[tokio::test]
-    async fn block_fee_read_and_increase_can_delete() {
-        let storage = cnidarium::TempStorage::new().await.unwrap();
-        let snapshot = storage.latest_snapshot();
-        let mut state = StateDelta::new(snapshot);
-
-        // can write
-        let asset_first = asset_0();
-        let asset_second = asset_1();
-        let amount_first = 100u128;
-        let amount_second = 200u128;
-
-        state
-            .get_and_increase_block_fees(&asset_first, amount_first, "test".into())
-            .await
-            .unwrap();
-        state
-            .get_and_increase_block_fees(&asset_second, amount_second, "test".into())
-            .await
-            .unwrap();
-        // holds expected
-        let fee_balances = HashSet::<_>::from_iter(state.get_block_fees().await.unwrap());
-        assert_eq!(
-            fee_balances,
-            HashSet::from_iter(vec![
-                (asset_first.to_ibc_prefixed(), amount_first),
-                (asset_second.to_ibc_prefixed(), amount_second)
-            ]),
-            "returned fee balance vector not what was expected"
-        );
-
-        // can delete
-        state.clear_block_fees().await;
-
-        let fee_balances_updated = state.get_block_fees().await.unwrap();
-        assert!(
-            fee_balances_updated.is_empty(),
-            "fee balances were expected to be deleted but were not"
-        );
-    }
-
-    #[tokio::test]
-    async fn is_allowed_fee_asset() {
-        let storage = cnidarium::TempStorage::new().await.unwrap();
-        let snapshot = storage.latest_snapshot();
-        let mut state = StateDelta::new(snapshot);
-
-        // non-existent fees assets return false
-        let asset = asset_0();
-        assert!(
-            !state
-                .is_allowed_fee_asset(&asset)
-                .await
-                .expect("checking for allowed fee asset should not fail"),
-            "fee asset was expected to return false"
-        );
-
-        // existent fee assets return true
-        state.put_allowed_fee_asset(&asset);
-        assert!(
-            state
-                .is_allowed_fee_asset(&asset)
-                .await
-                .expect("checking for allowed fee asset should not fail"),
-            "fee asset was expected to be allowed"
-        );
-    }
-
-    #[tokio::test]
-    async fn can_delete_allowed_fee_assets_simple() {
-        let storage = cnidarium::TempStorage::new().await.unwrap();
-        let snapshot = storage.latest_snapshot();
-        let mut state = StateDelta::new(snapshot);
-
-        // setup fee asset
-        let asset = asset_0();
-        state.put_allowed_fee_asset(&asset);
-        assert!(
-            state
-                .is_allowed_fee_asset(&asset)
-                .await
-                .expect("checking for allowed fee asset should not fail"),
-            "fee asset was expected to be allowed"
-        );
-
-        // see can get fee asset
-        let assets = state.get_allowed_fee_assets().await.unwrap();
-        assert_eq!(
-            assets,
-            vec![asset.to_ibc_prefixed()],
-            "expected returned allowed fee assets to match what was written in"
-        );
-
-        // can delete
-        state.delete_allowed_fee_asset(&asset);
-
-        // see is deleted
-        let assets = state.get_allowed_fee_assets().await.unwrap();
-        assert!(assets.is_empty(), "fee assets should be empty post delete");
-    }
-
-    #[tokio::test]
-    async fn can_delete_allowed_fee_assets_complex() {
-        let storage = cnidarium::TempStorage::new().await.unwrap();
-        let snapshot = storage.latest_snapshot();
-        let mut state = StateDelta::new(snapshot);
-
-        // setup fee assets
-        let asset_first = asset_0();
-        state.put_allowed_fee_asset(&asset_first);
-        assert!(
-            state
-                .is_allowed_fee_asset(&asset_first)
-                .await
-                .expect("checking for allowed fee asset should not fail"),
-            "fee asset was expected to be allowed"
-        );
-        let asset_second = asset_1();
-        state.put_allowed_fee_asset(&asset_second);
-        assert!(
-            state
-                .is_allowed_fee_asset(&asset_second)
-                .await
-                .expect("checking for allowed fee asset should not fail"),
-            "fee asset was expected to be allowed"
-        );
-        let asset_third = asset_2();
-        state.put_allowed_fee_asset(&asset_third);
-        assert!(
-            state
-                .is_allowed_fee_asset(&asset_third)
-                .await
-                .expect("checking for allowed fee asset should not fail"),
-            "fee asset was expected to be allowed"
-        );
-
-        // can delete
-        state.delete_allowed_fee_asset(&asset_second);
-
-        // see is deleted
-        let assets = HashSet::<_>::from_iter(state.get_allowed_fee_assets().await.unwrap());
-        assert_eq!(
-            assets,
-            HashSet::from_iter(vec![
-                asset_first.to_ibc_prefixed(),
-                asset_third.to_ibc_prefixed()
-            ]),
-            "delete for allowed fee asset did not behave as expected"
-        );
-    }
-
-    #[test]
-    fn storage_keys_are_not_changed() {
-        let trace_prefixed = "a/denom/with/a/prefix"
-            .parse::<astria_core::primitive::v1::asset::Denom>()
-            .unwrap();
-        assert_eq!(
-            block_fees_key(&trace_prefixed),
-            block_fees_key(trace_prefixed.to_ibc_prefixed()),
-        );
-        insta::assert_snapshot!(block_fees_key(&trace_prefixed));
-
-        assert_eq!(
-            fee_asset_key(&trace_prefixed),
-            fee_asset_key(trace_prefixed.to_ibc_prefixed()),
-        );
-        insta::assert_snapshot!(fee_asset_key(trace_prefixed));
     }
 }
